backend/requirements.txt
fastapi[standard]>=0.115
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
pyjwt>=2.0
pwdlib[argon2]>=0.2
pydantic-settings>=2.0
python-multipart>=0.0.9
uvicorn>=0.30
APScheduler==3.11.2
azure-mgmt-costmanagement
azure-identity
tenacity
anthropic>=0.40
redis>=5.0
aiosmtplib>=3.0
httpx>=0.27
